+++ packages/file-manager/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @cmdoss/file-manager
 
-## 2.0.0
+## 1.2.0
 
 ### Minor Changes
 
@@ -9,7 +9,7 @@
 ### Patch Changes
 
 - Updated dependencies [a4e8041]
-  - @cmdoss/site-builder@2.0.0
+  - @cmdoss/site-builder@1.2.0
 
 ## 1.0.1