+++ packages/site-builder/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @cmdoss/site-builder
 
-## 2.0.0
+## 1.2.0
 
 ### Minor Changes